+++ R/record_versions.R
@@ -12,18 +12,14 @@
 #' @param manifest Character of length 1, file path to the JSON manifest.
 #' @param issues Character of length 1, file path to a JSON file
 #'   which records packages with version issues.
-#' @param repos Character string of package repositories to track.
+#' @param repo Character string of package repositories to track.
 #' @param current A data frame of current versions and hashes of packages
-#'   in `repos`. This argument is exposed for testing only.
-#' @param check_hash Logical of length 1, check hashes when judging package
-#'   version compliance. This allows [record_versions()] to flag packages
-#'   that create new releases but keep the same version number.
+#'   in `repo`. This argument is exposed for testing only.
 record_versions <- function(
   manifest = "versions.json",
   issues = "version_issues.json",
-  repos = "https://r-releases.r-universe.dev",
-  current = r.releases.utils::get_current_versions(repos = repos),
-  check_hash = TRUE
+  repo = "https://r-releases.r-universe.dev",
+  current = r.releases.internals::get_current_versions(repo = repo)
 ) {
   if (!file.exists(manifest)) {
     jsonlite::write_json(x = current, path = manifest, pretty = TRUE)
@@ -32,7 +28,7 @@ record_versions <- function(
   previous <- read_versions_previous(manifest = manifest)
   new <- update_version_manifest(current = current, previous = previous)
   jsonlite::write_json(x = new, path = manifest, pretty = TRUE)
-  aligned <- versions_aligned(manifest = new, check_hash = check_hash)
+  aligned <- versions_aligned(manifest = new)
   new_issues <- new[!aligned,, drop = FALSE] # nolint
   jsonlite::write_json(x = new_issues, path = issues, pretty = TRUE)
   invisible()
@@ -41,15 +37,25 @@ record_versions <- function(
 #' @title Get the current versions of packages
 #' @export
 #' @keywords internal
-#' @description Get the current versions of packages in the repos.
+#' @description Get the current versions of packages in the repo.
 #' @return A data frame of packages with their current versions and hashes.
 #' @inheritParams record_versions
 get_current_versions <- function(
-  repos = "https://r-releases.r-universe.dev"
+  repo = "https://r-releases.r-universe.dev"
 ) {
-  out <- utils::available.packages(repos = repos)
-  out <- out[, c("Package", "Version", "MD5sum")]
-  out <- as.data.frame(out)
+  listing <- paste(
+    contrib.url(repos = repo, type = "source"),
+    "PACKAGES.json?fields=RemoteSha",
+    sep = "/"
+  )
+  out <- jsonlite::stream_in(
+    con = url(listing),
+    verbose = TRUE,
+    simplifyVector = TRUE,
+    simplifyDataFrame = TRUE,
+    simplifyMatrix = TRUE
+  )
+  out <- out[, c("Package", "Version", "RemoteSha")]
   colnames(out) <- c("package", "version_current", "hash_current")
   rownames(out) <- NULL
   out
@@ -91,10 +97,7 @@ manifest_compare_versions <- function(manifest) {
   )
 }
 
-versions_aligned <- function(manifest, check_hash) {
-  aligned <- manifest$version_current == manifest$version_highest
-  if (check_hash) {
-    aligned <- aligned & (manifest$hash_current == manifest$hash_highest)
-  }
-  aligned
+versions_aligned <- function(manifest) {
+  (manifest$version_current == manifest$version_highest) &
+    (manifest$hash_current == manifest$hash_highest)
 }