+++ tests/test-record_versions.R
@@ -13,7 +13,7 @@ contents <- data.frame(
   version_current = rep("1.0.0", 4L),
   hash_current = rep("hash_1.0.0", 4L)
 )
-r.releases.utils::record_versions(
+r.releases.internals::record_versions(
   manifest = manifest,
   issues = issues,
   current = contents
@@ -45,7 +45,7 @@ stopifnot(identical(written, expected))
 stopifnot(!file.exists(issues))
 
 # Update the manifest after no changes to packages or versions.
-r.releases.utils::record_versions(
+r.releases.internals::record_versions(
   manifest = manifest,
   issues = issues,
   current = contents
@@ -96,7 +96,7 @@ index <- contents$package == "version_unmodified"
 contents$version_current[index] <- "1.0.0"
 contents$hash_current[index] <- "hash_1.0.0-modified"
 for (index in seq_len(2L)) {
-  r.releases.utils::record_versions(
+  r.releases.internals::record_versions(
     manifest = manifest,
     issues = issues,
     current = contents
@@ -154,67 +154,15 @@ for (index in seq_len(2L)) {
   stopifnot(identical(written_issues, expected_issues))
 }
 
-# Same, but do not check the hash.
-r.releases.utils::record_versions(
-  manifest = manifest,
-  issues = issues,
-  current = contents,
-  check_hash = FALSE
-)
-written <- jsonlite::read_json(manifest)
-expected <- list(
-  list(
-    package = "package_unmodified",
-    version_current = "1.0.0",
-    hash_current = "hash_1.0.0",
-    version_highest = "1.0.0",
-    hash_highest = "hash_1.0.0"
-  ),
-  list(
-    package = "version_decremented",
-    version_current = "0.0.1",
-    hash_current = "hash_0.0.1",
-    version_highest = "1.0.0",
-    hash_highest = "hash_1.0.0"
-  ),
-  list(
-    package = "version_incremented",
-    version_current = "2.0.0",
-    hash_current = "hash_2.0.0",
-    version_highest = "2.0.0",
-    hash_highest = "hash_2.0.0"
-  ),
-  list(
-    package = "version_unmodified",
-    version_current = "1.0.0",
-    hash_current = "hash_1.0.0-modified",
-    version_highest = "1.0.0",
-    hash_highest = "hash_1.0.0"
-  )
-)
-stopifnot(identical(written, expected))
-stopifnot(file.exists(issues))
-written_issues <- jsonlite::read_json(issues)
-expected_issues <- list(
-  list(
-    package = "version_decremented",
-    version_current = "0.0.1",
-    hash_current = "hash_0.0.1",
-    version_highest = "1.0.0",
-    hash_highest = "hash_1.0.0"
-  )
-)
-stopifnot(identical(written_issues, expected_issues))
-
 # Remove temporary files
 unlink(c(manifest, issues))
 
 # The manifest can be created and updated from the actual repo.
 manifest <- tempfile()
 issues <- tempfile()
-r.releases.utils::record_versions(manifest = manifest, issues = issues)
+r.releases.internals::record_versions(manifest = manifest, issues = issues)
 stopifnot(file.exists(manifest))
-r.releases.utils::record_versions(manifest = manifest, issues = issues)
+r.releases.internals::record_versions(manifest = manifest, issues = issues)
 contents <- jsonlite::read_json(manifest)
 stopifnot(is.character(contents[[1L]]$package))
 stopifnot(length(contents[[1L]]$package) == 1L)